PLUGINS-3132 Enhance Jenkins Plugin for Velocity
Salesforce case for the Velocity Jenkins plugin- TS004660800. The 'Upload Deployment' step wasn't working as expected for the customer. It looks like versionName and versionExternal ID was not filled by the customer and hence it was failing. The customer is using a free style Jenkins job and using the UI option to configure the upload Deployment step. The versionName and versionExternal ID are in Advanced options and he suggested moving it outside.